requirements.txt
langchain
langchain-community
langchain-openai
langchain-core
langchain-experimental
langchain-google-genai

streamlit
pandas
numpy
matplotlib
seaborn
pygwalker
jupyterlab
tiktoken
tabulate
python-dotenv
pygwalker

